refactor: Name 클래스 생성
- 이름 길이에 대한 검증 책임을 Player -> Name 으로 옮김
