+++ src/main/java/view/OutputView.java
@@ -3,6 +3,7 @@ package view;
 import java.util.List;
 
 import domain.Line;
+import domain.Name;
 import domain.Player;
 
 public class OutputView {
@@ -55,7 +56,7 @@ public class OutputView {
     }
 
     private String getNameBlank(String name) {
-        return BLANK.repeat(Player.NAME_MAX_LENGTH - name.length());
+        return BLANK.repeat(Name.MAX_LENGTH - name.length());
     }
 
 }
